move factory references
move validator reference
